requirements.txt
acconeer-exptool==7.16.0
attributes-doc==0.4.0
attrs==25.3.0
cbor2==5.6.5
certifi==2025.6.15
charset-normalizer==3.4.2
commonmark==0.9.1
exceptiongroup==1.1.1
h5py==3.14.0
idna==3.10
numpy==2.3.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.0
platformdirs==4.3.8
Pygments==2.19.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
PyYAML==6.0.2
requests==2.32.4
rich==12.6.0
six==1.17.0
smbus2==0.5.0
textual==0.1.18
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
xlsxwriter==3.2.5
